[Issue #134] Changed Type of Merch Price from int to Double

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/dto/MerchandiseDto.java
@@ -14,8 +14,8 @@ public class MerchandiseDto {
     private String merchandiseProductName;
     private int stockCount;
     private String photo;
-    private int price;
-    private int premiumPrice;
+    private Double price;
+    private Long premiumPrice;
     private boolean premium;
 
     @Override
